% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.demogdata.R
\name{as.data.frame.demogdata}
\alias{as.data.frame.demogdata}
\title{Coerce a demogdata object to a data.frame object}
\usage{
\method{as.data.frame}{demogdata}(x, ...)
}
\arguments{
\item{x}{Object to be coerced to a data frame.}

\item{...}{Other arguments not used}
}
\value{
A data.frame object.
}
\description{
Coerce a demogdata object to a data.frame object
}
\examples{
# coerce demogdata object to data.frame ----
as.data.frame(fr.mort)
}
