% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogdata.R
\name{sex.ratio}
\alias{sex.ratio}
\title{Compute sex ratios from mortality rates}
\usage{
sex.ratio(data)
}
\arguments{
\item{data}{Demogdata object of type \dQuote{mortality} such as obtained from \code{\link{read.demogdata}},
or an object of class \code{fmforecast} such as the output from  \code{\link{forecast.fdm}} or \code{\link{forecast.lca}}.}
}
\value{
Functional time series of sex ratios.
}
\description{
Calculates the Male/Female ratios from historical or forecasted mortality rates.
}
\examples{
plot(sex.ratio(fr.mort),ylab="Sex ratios (M/F)")
}
\author{
Rob J Hyndman
}
\keyword{models}
