\name{demography-package}
\alias{demography-package}
\alias{demography}
\docType{package}
\title{Forecasting mortality and fertility data}
\description{Functions for demographic analysis including lifetable calculations,
             Lee-Carter modelling and functional data analysis of mortality rates.
}

\author{
Rob J Hyndman, Heather Booth, Leonie Tickle, John Maindonald

Maintainer: <Rob.Hyndman@buseco.monash.edu.au>
}

\keyword{ package }

