% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography.R
\docType{package}
\name{demography-package}
\alias{demography-package}
\alias{demography}
\title{Forecasting mortality and fertility data}
\description{
Functions for demographic analysis including lifetable calculations,
Lee-Carter modelling and functional data analysis of mortality rates.
}
\author{
Rob J Hyndman with contributions from Heather Booth, Leonie Tickle, John Maindonald, Simon Wood and the R Core Team.

Maintainer: <Rob.Hyndman@monash.edu>
}
\keyword{package}
