% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogdata.R
\name{extract.years}
\alias{extract.years}
\title{Extract some years from a demogdata object}
\usage{
extract.years(data, years)
}
\arguments{
\item{data}{Demogdata object such as created using \code{\link{read.demogdata}} or \code{\link{smooth.demogdata}}.}

\item{years}{Vector of years to extract from data.}
}
\value{
Demogdata object with same components as \code{data} but with a subset of years.
}
\description{
Creates subset of demogdata object.
}
\examples{
france.1918 <- extract.years(fr.mort,1918)
}
\author{
Rob J Hyndman
}
\keyword{manip}
