\name{median}
\alias{median}
\alias{median.default}
\title{Median}
\usage{median(...)
}

\arguments{
\item{...}{Arguments passed to specific methods.}
}

\description{Generic function for the median.}

\details{The \code{\link[stats]{median}} function in the \code{stats} package is replaced by \code{median.default}.}

\value{Refer to specific methods. For numeric vectors, see the \code{\link[stats]{median}} function in the \code{stats} package.}

\author{Rob J Hyndman}

\seealso{\code{\link{median.demogdata}}}

\keyword{models}
