% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogdata.R
\name{set.upperage}
\alias{set.upperage}
\title{Combine the upperages of a demogdata object.}
\usage{
set.upperage(data, max.age)
}
\arguments{
\item{data}{Demogdata object such as created using \code{\link{read.demogdata}} or \code{\link{smooth.demogdata}}.}

\item{max.age}{Upper age group. Ages beyond this are combined into the upper age group.}
}
\value{
Demogdata object with same components as \code{data} but with a subset of ages.
}
\description{
Computes demographic rates by combining age groups.
}
\examples{
france.short <- set.upperage(fr.mort, 85)
}
\author{
Rob J Hyndman
}
\keyword{manip}
